+++ code/programs/mosaic/venture-browser/host/flutter/mosaic_host.dart
@@ -292,6 +292,7 @@ class MosaicHost {
   Map<String, Object?>? lastBrowsingContextRequest;
   Map<String, Object?>? lastDownloadRequest;
   Map<String, Object?>? lastPrintRequest;
+  Map<String, Object?>? lastShareRequest;
   String? lastClipboardText;
   Stream<Map<String, Object?>> get filePickerRequests => _filePickerRequests.stream;
 
@@ -528,6 +529,8 @@ class MosaicHost {
       lastDownloadRequest = Map<String, Object?>.from(effect);
     } else if (effect['type'] == 'print') {
       lastPrintRequest = Map<String, Object?>.from(effect);
+    } else if (effect['type'] == 'share') {
+      lastShareRequest = Map<String, Object?>.from(effect);
     } else if (effect['type'] == 'write-clipboard' && effect['text'] is String) {
       lastClipboardText = effect['text'] as String;
       unawaited(Clipboard.setData(ClipboardData(text: lastClipboardText!)));